Scans a DICOM directory and extracts information about each series. The files in the directory must conform to the following naming convention:
        NNNNNN-S-MMMMM.dcm

where NNNNNN and MMMMM are any strings of any length as long as thay
do not contain dashes (-). S is the series/run number. This is the
naming convention of the data at MGH after being pushed from the scanner
to the network archive. This function will not work on data burned to
CD from the scanner.  Example: 953000-2-9.dcm