FS-FAST routine to mask out regions from a contrast that are not in the brain. You can specify multiple contrasts and multiple maps inside a contrast (all maps must exist inside all contrasts).
The routine will use the brain mask as found in bold/masks. This
can be created with mkbrainmask-sess. All voxels outside of this
region will be set to 0. The resulting map will be called map-brain
and can be accessed with all the commands that would normally be
used to viewed maps (eg, sliceview, tkmedit, etc).