SUMMARY
Converts a label or a set of labels into a volume. For a single lable,
the volume will be binary: 1 where the label is and 0 where it is not.
For multiple labels, the volume will be 0 where no labels were found
otherwise the value will the the label number. For a voxel to be
declared part of a label, it must have enough hits in the voxel and
it must have more hits than any other label.