o MRIScomputeSecondFundamentalForm(...) o MRIScomputeSecondFundamentalFormAtVertex(...)
For ill-conditioned events at a vertex when computing the
second order fundamental form, individual K and H curvatures
can greatly exceed the current surface Kmax and Hmax, producing
large outliers. This skews statistics on these surfaces.

The correction limits the vertex K and H to, in this case,
the current surface Kmax and Hmax.