Rewritten to solve problems between multiple indexing schemes used in this code, between assuming that tables had no empty or skipped fields. New code allows NULL or empty entries, but clients must be careful that they are thinking of 'structure indecies' in the right way.
- Change the color table so that we don't have to convert structure to
index; it is costly as it involves a O(n) search through the
table. The table in memory will just have empty entries for the
skipped structures. However, it means that our color tables become
crazy long when we skip a lot of entries (as in the
FreeSurferColorLUT.txt which goes up to 5000 entries), so we waste a
lot of memory. (All my code is doing this internally already, tho, as
it's too slow to do it the CTAB way; this is why CTAB isn't breaking
tkmedit and scuba right now.)

- Change all the CTAB functions so they only use 'index', which means
both 0-based table index as well as structure number.

- Change the binary CTAB format so that it writes a negative version
number first. Then, when we read the first int from the binary stream,
we see if it's < 0, and if so, use the new version 2 format.

- The version 2 format of the binary CTAB code will write structure
numbers so we our binary format is in line with the ASCII file format.