fscalc.fsl - a freesurfer wrapper around fslmaths
This is a utility that extends fslmaths so that any format that can
be read/written by freesurfer can be used. It operates in the same
way as fslmaths, just use the same command line flags and arguments
with the exception that the input file formats can be anything that
freesurfer reads, and the output file format can be anything that
freesurfer writes. Note that volume-encoded surface files will still
fail due to nifti dimension incompatibility. See fslmaths for more
info.