Copied this change from dev:
Modified computation of cortical volume so that it did not interfere
with the ocmputation of white matter volume. Still not sure that the
cortical volume computation is correct (does it include non-cortical
parts of the surface?). Aslo now prints out the total wm volume (l+r)
and total ctx volume.