added support for embedded colortables
added functions for setting voxels with a given value to a different
value in an output volume.

added functions for setting voxels with a given value in one volume to
another value in a different volume.

program for compiling all edits in a subject's mri dir into a single
segmentation volume.