Performs a functional-structural registration.
By default you only need to supply the sessions and register-sess will
perform the registration. The default is to use bbregister initialized
with fsl (FLIRT) and to create a file called register.dat in the bold
directory. Each session must have a subjectname file with the freesurfer
subject name in it.