Added more robust checking for whether inputs are valid files or float arguments.
'mris_calc' was susceptible to being fooled if input file arg spec
started with a sequence of numeric chars. The test for conversion
of name to numeric failed, and in the process corrupted the file
arg, resulting in 'mris_calc' not being able to find the file
either.

This now more robustly handled by 'mris_calc' checking if input
arg is a valid file using stat(...) -- if file does not exist,
'mris_calc' then attempts to convert to float. Float conversion
is only considered valid if no spurious characters remain at
end of conversion string. If spurious characters remain (as in
a strtof conversion of, for e.g., '2.145test' then the string
is assumed to refer to a file, which will then fail.

If a filespec is masked by a file on the filesystem with the same
name, then the file is interpreted and NOT the float conversion.
For example, if you have a file called '100' which a valid mgz
volume, and run 'mris_calc somefile.mgz mul 100', the the file
'100' will be accessed and not considered to mean the number
100.