Add two new stats outputs:
zeroCount	- number of "zeros" in argument volume/overlay
nonZeroCount	- number of "!zeros" in argument volume/overlay

Note: These are not printed to "PrintStatToFile". I'm not sure what
the purpose of that function is, so I don't want to change any
downstream behaviour that depends on a strict ordering of stats
in that file.