This is an atlas that includes sulcal CSF, skull/bone, head tissue, and air cavities insdie the head as well as pons and vermis.  It was generated from 76 subjects from the IXI data set http://www.brain-development.org (see below for demographics).
The extra cerebral labels were generated by first analyzing the T1s in
SPM "NewSeg" using the spmseg script. NewSeg creates segmentations
that include sulcal CSF, bone, and air. This seg was modified by
replacing the brain voxels with those from the aseg.

The vermis segmentation was complicated. The Buckner 40 was analyzed
in Suit to generate a vermis segmentation. This was then used to train
a gca. This gca was applied to the IXI data. It was then included
in the extra cerebral segmentation above.

The pons segmentation was performed by first hand labeling pons on the
MNI152. This was mapped to each IXI subject through a linear transform
and intersected with the brainstem label.

These segmentations where then used to train the atlas.

ntotal 79
males 34
females 45
white 39
asian 14
black 13
chinese 13
Guys-Philips-1.5T 39
HH-Philips-3T 22
IOP-GE-1.5T 18
mean age: 58.0995
min age: 20.21
max age: 86.32
 20-30  5
 30-40 10
 40-50 10
 50-60 13
 60-70 17
 70-80 16
 80-90  8