Uupdate to build latest vxl code from git.  No recent tag would build even with gcc 5.5+ or cmake 3.3+, so use that hash of the last commit as the checkout tag (from 4/18/2018).
Add defines for compilers and cmake on Ubuntu,
centOS7, and Mac OS. If compiler cannot be found,
then don't build anything and do not fail the
build. I did not see that anything links against
the built libs, so this should be OK for now.

The latest vxl source does build with gcc5.5
on Mac, gcc 6.3 on CentOS 7, and the default
tools on Ubuntu 16.04 - see comments in
makefile for more details.