Fixing possible infinite recursion in trc/blood.cxx.
FindCenterStreamline function (in trc/blood.cxx) calls itself with different checking options set to true or false in certain cases. The base case (else branch on line 2429) potentially calls itself to infinity if it happens that hdmin's value is not changed from numeric_limits<double>::infinity(). I presume this has to do with some characteristic of a streamline being unsatisfactory. In any case, FindCenterStreamline should not call itself again if all the check variables are already set to false.