Added defect_labels to CommandLoadSubject
CommandLoadSubject now automatically loads the ?h.defect_labels onto orig.nofix surfaces to QA. The labels are visible in the 2D displays thanks to the edgecolor=overlay option. I've included a threshold of 0.01% -- 100% to include all defect labels and set the visibility to 0 in case the user wants the surface out of sight.