Add CNN segmentation of thalamic nuclei from DTI
Add python code, supporting .npy files and modify the FreeSurferColorLUT
in preparation for a CNN implementation of a thalamic nuclei segmentation
for joint T1 and DTI volumes.

This commit includes modifications to the CMakeLists that are required
to build FreeSurfer and this new package.

This commit does not include the required model weights that need to be
uploaded to the data annex.