bf: large FOV memory bad alloc
Some volumes default to an in plane field of view of 256 by 256 voxels
even when their voxel sizes are larger than 1mm in plane. This can
result in much larger tensors when pre-upsampling to 1mm isotropic,
which in turn can cause unhandled memory allocation issues on smaller
machines.

The fix here it to include an --autocrop option which detects non-
background voxels and crops to a bounding box. Here non-background is
defined as voxel intensity greater than 0 after the re-normalisation
and intensity clipping steps. Testing on ADNI led to a greater than
fourfold reduction in tensor size.